cell_genes_7 {SCPA}	R Documentation

Cell genes

Description

Selection and enrichment analysis of differential genes in cells.

Usage


cell_genes_7(
  sce,
  group = "annotation",
  group_cell_1 = "B_cell",
  group_cell_2 = "NK_cell",
  Two_cell_enrichment = TRUE,
  dir = "./6.Cell genes"
)

Arguments

sce
Single cell Seurat data.

group
Grouping data of single cells.

group_cell_1
Select cell type 1.

group_cell_2
Select cell type 2.

Two_cell_enrichment
Analysis of differential gene enrichment in two types of cells.

dir
Output data directory.

Value

a sce object

[Package SCPA version 1.0.1 Index]