Create pathway matrices from gene sets

This function takes a matrix and pathway list as and input and creates expression matrices for each pathway. The resulting output is a list of expression matrices for each pathway

pathway_matrices(samples, pathways, sample_names = NULL, pathway_names = NULL)

Arguments

samples
List containing expression file of samples

pathways
List of pathways to create pathway specific matrices. You can use the get_paths function to create this list if you are using a csv file generated from a gmt file.

sample_names
Names of the samples to be used in output

pathway_names
Names of pathways to be used in output