Plot a heatmap of qvals from SCPA

This function takes the output from SCPA and plots a heatmap of the qvals

plot_heatmap(
  scpa_out,
  highlight_pathways = NULL,
  row_fontsize = 6,
  column_fontsize = 10,
  column_names = colnames(scpa_out),
  show_row_names = TRUE,
  cluster_columns = TRUE,
  hm_colors = NULL,
  scale_breaks = NULL
)

Arguments

scpa_out

Data frame that contains a "Pathway" column, and a column or multiple columns containing qvals. This can be the direct output from compare_pathways(), or a custom data frame.

highlight_pathways

Pathway or pathways to annotate on the heatmap, supplied as character vector. If no argument is given, all pathways are shown

row_fontsize

Font size of pathway names

column_fontsize

Font size of the sample names

column_names

Option to supply names of the heatmap columns if more than one populations is present

show_row_names

Should row names be shown in the heatmap?

cluster_columns

Should columns in the heatmap be clustered?

hm_colors

Colors to be used in the heatmap

scale_breaks

Breaks to be used in the colors of the heatmap. Length must be equal to the number of colors.

Value

list of pathways with corresponding genes

Examples

if (FALSE) {
plot_heatmap(
     scpa_out = scpa_result,
     pathway = "mtorc"
)
}