Use SCPA to compare gene sets

This function takes an input of samples and pathways to compare gene sets over different conditions.

compare_pathways(
  samples,
  pathways,
  downsample = 500,
  min_genes = 15,
  max_genes = 500
)

Arguments

samples
List of samples, each supplied as an expression matrix

pathways
List of pathways and their genes

downsample
Option to downsample cell numbers. Default is 500

min_genes
Gene sets with fewer than this number of genes will be excluded

max_genes
Gene sets with more than this number of genes will be excluded

Value

Statistical results from the SCPA analysis. If only two samples are provided, an enrichment score will also be calculated.

Examples

if (FALSE) {
scpa_result <- compare_pathways(
     list(sample1, sample2, sample3),
     pathways = pathways)
}