Plot the rank of specific pathways from SCPA output

This function takes the output from SCPA and plots the rank of a user defined pathway.

plot_rank(
  scpa_out,
  pathway,
  population_name = "qval",
  base_point_size = 2,
  base_point_color = "gray70",
  highlight_point_size = 3,
  highlight_point_color = "cornflowerblue",
  label_pathway = T,
  label_size = 4
)

Arguments

scpa_out

Data frame containing Pathways and qvals generated from compare_pathways

pathway

Chosen pathway or pathways to plot the rank of. This can be specific e.g. HALLMARK_GLYCOLYSIS to plot a specific result, or generic e.g. glycolysis to plot all glycolysis pathways

population_name

Column name of the population to plot, if more than one population present in the data frame

base_point_size

Size of the base points to plot on the graph

base_point_color

Color of the base points to plot on the graph

highlight_point_size

Size of the highlighted points to plot on the graph

highlight_point_color

Color of the highlighted points to plot on the graph

label_pathway

Should the selected pathway be labelled?

label_size

Text size of the pathway label

Value

list of pathways with corresponding genes

Examples

if (FALSE) {
plot_rank(
     scpa_out = scpa_result,
     pathway = "interferon",
     population_name = cd4_qval,
)
}